Module Tezos_raw_protocol_007_PsDELPH1.Roll_storage

Basic roll manipulation.

If storage related to roll (i.e. `Storage.Roll`) is not used outside of this module, this interface enforces the invariant that a roll is always either in the limbo list or in a contract list.

type Tezos_protocol_environment_007_PsDELPH1__Environment.Error_monad.error +=

| Consume_roll_change

| No_roll_for_delegate

| No_roll_snapshot_for_cycle of Cycle_repr.t

| Unregistered_delegate of Tezos_crypto.Signature.public_key_hash

val init : Raw_context.t -> (Raw_context.t, Tezos_protocol_environment_007_PsDELPH1__Environment.Error_monad.error list) Stdlib.result Lwt.t

val init_first_cycles : Raw_context.t -> (Raw_context.t, Tezos_protocol_environment_007_PsDELPH1__Environment.Error_monad.error list) Stdlib.result Lwt.t

val cycle_end : Raw_context.t -> Cycle_repr.t -> (Raw_context.t, Tezos_protocol_environment_007_PsDELPH1__Environment.Error_monad.error list) Stdlib.result Lwt.t

val snapshot_rolls : Raw_context.t -> (Raw_context.t, Tezos_protocol_environment_007_PsDELPH1__Environment.Error_monad.error list) Stdlib.result Lwt.t

val fold : Raw_context.t -> f:(Roll_repr.roll -> Tezos_crypto.Signature.public_key -> 'a -> ('a, Tezos_protocol_environment_007_PsDELPH1__Environment.Error_monad.error list) Stdlib.result Lwt.t) -> 'a -> ('a, Tezos_protocol_environment_007_PsDELPH1__Environment.Error_monad.error list) Stdlib.result Lwt.t

val baking_rights_owner : Raw_context.t -> Level_repr.t -> priority:int -> (Tezos_crypto.Signature.public_key, Tezos_protocol_environment_007_PsDELPH1__Environment.Error_monad.error list) Stdlib.result Lwt.t

val endorsement_rights_owner : Raw_context.t -> Level_repr.t -> slot:int -> (Tezos_crypto.Signature.public_key, Tezos_protocol_environment_007_PsDELPH1__Environment.Error_monad.error list) Stdlib.result Lwt.t

module Delegate : sig ... end

module Contract : sig ... end

val delegate_pubkey : Raw_context.t -> Tezos_crypto.Signature.public_key_hash -> (Tezos_crypto.Signature.public_key, Tezos_protocol_environment_007_PsDELPH1__Environment.Error_monad.error list) Stdlib.result Lwt.t

val get_rolls : Raw_context.t -> Tezos_crypto.Signature.public_key_hash -> (Roll_repr.t list, Tezos_protocol_environment_007_PsDELPH1__Environment.Error_monad.error list) Stdlib.result Lwt.t

val get_change : Raw_context.t -> Tezos_crypto.Signature.public_key_hash -> (Tez_repr.t, Tezos_protocol_environment_007_PsDELPH1__Environment.Error_monad.error list) Stdlib.result Lwt.t

val update_tokens_per_roll : Raw_context.t -> Tez_repr.t -> (Raw_context.t, Tezos_protocol_environment_007_PsDELPH1__Environment.Error_monad.error list) Stdlib.result Lwt.t